Remove latest changes
The latest changes are substanstial and have been made by community
members. While the help is very appreciated i don't have much time
to spend on the lib, so was not able to verify that those changes
are not breaking things. So in order to be safe and be able to
relase updates to the library i decided the easies way to move
forward is to remove the changes. As i don't have time to make sure
they are safe.