Rename full screen logic to be more explicit
The existing methods to handle full screen mode in `YouTubePlayerView` only change the view’s `LayoutParams` to fill its parent or wrap its content.
The naming is misleading because they don’t change the state of the IFrame Player. This change is in preparation of future integration with the full screen state of the IFrame Player.

List of main changes in this commit (these are also breaking changes):

* Replace YouTubePlayerView#enterFullScreen with YouTubePlayerView#matchParent
* Replace YouTubePlayerView#existFullScreen with YouTubePlayerView#wrapContent
* Remove YouTubePlayerView#toggleFullScreen
* Remove YouTubePlayerView#isFullScreen
* Remove `YouTubePlayerFullScreenListener` interface