Remove youtube.com from default origin
As discussed in #1252, the default origin can't be "https://www.youtube.com" anymore. With this origin the IFrame Player API will reject the request and fail to play the video.

According to the updated IFrame player documentation, the origin should ideally be set to "https://[app-id]", where app-id is the id of the app.

From within the library we don't know the app-id of the client app, therefore the default origin value is now set to "https://com.pierfrancescosoffritti.androidyoutubeplayer".

This should make no different from YouTube point's of view. Client can set their own app-id by providing an IFramePlayerOptions to the YouTubePlayerView initialization method.